refactor:SkeltonComponent
created a seperate component for skelton
