+++ Lecture11/src/components/NoRecordComponent.js
@@ -1,6 +1,6 @@
 const NoRecord = () => (
     <div className="noRecord">
-        <h3>No Record Found</h3>
+        <h3 className="font-bold text-2xl text-white">No Record Found</h3>
     </div>
 );
